pyproject.toml
[project]
name = "nautilus-binance-v2"
version = "0.1.0"
description = "Institutional-grade crypto trading system built on NautilusTrader"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Leo" }]

dependencies = [
    "nautilus_trader>=1.223.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "prometheus-client>=0.24.1",
    "structlog>=25.5.0",
    "tqdm>=4.0",
    "psycopg[binary]>=3.2",
    "redis[hiredis]>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.0",
    "pre-commit>=4.0",
]
alerting = [
    "python-telegram-bot>=21.0",
    "slack-sdk>=3.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "N", "B", "A", "C4", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.13"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.2",
]
